requirements.txt
astroid==2.9.3
colorama==0.4.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
pip==22.0.2
pylint==2.12.2
setuptools==60.7.0
six==1.16.0
typed-ast==1.52
wrapt==1.13.3
yapf==0.32.0
configparser==5.2.0